[Matrix]: fix some minor warning in dense_matrix.

+++ matrix/dense_matrix.cpp
@@ -2177,7 +2177,7 @@ vector::ptr dense_matrix::get_row(off_t idx) const
 dense_matrix::ptr dense_matrix::get_cols(const std::vector<off_t> &idxs) const
 {
 	for (size_t i = 0; i < idxs.size(); i++)
-		if (idxs[i] < 0 || idxs[i] >= get_num_cols()) {
+		if (idxs[i] < 0 || (size_t) idxs[i] >= get_num_cols()) {
 			BOOST_LOG_TRIVIAL(error) << "the col index is out of bound";
 			return dense_matrix::ptr();
 		}
@@ -2192,7 +2192,7 @@ dense_matrix::ptr dense_matrix::get_cols(const std::vector<off_t> &idxs) const
 dense_matrix::ptr dense_matrix::get_rows(const std::vector<off_t> &idxs) const
 {
 	for (size_t i = 0; i < idxs.size(); i++)
-		if (idxs[i] < 0 || idxs[i] >= get_num_rows()) {
+		if (idxs[i] < 0 || (size_t) idxs[i] >= get_num_rows()) {
 			BOOST_LOG_TRIVIAL(error) << "the row index is out of bound";
 			return dense_matrix::ptr();
 		}
